refactor: split tree shaker CJS patterns
